MyCollegeBot/requirements.txt
aiogram>=3.0
pydantic-settings
sqlalchemy
aiosqlite
python-dotenv